requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cachelib==0.10.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
Django==4.2.3
django-cors-headers==4.1.0
django-simple-history==3.3.0
djangorestframework==3.14.0
dnspython==2.1.0
et-xmlfile==1.1.0
Flask==2.2.3
Flask-Session==0.4.0
ftfy==6.1.1
gunicorn==20.1.0
httplib2==0.20.2
hyperlink==21.0.0
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.9.1
Markdown==3.3.6
MarkupSafe==2.1.2
more-itertools==8.10.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
openpyxl==3.1.2
pexpect==4.8.0
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-magic==0.4.24
pytz==2023.3
redis==4.5.4
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sqlparse==0.4.4
ssh-import-id==5.11
Twisted==22.1.0
typing_extensions==4.7.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==1.0.0
zope.interface==5.4.0
